move task implementation out of 'package main'
this will make the behavior a lot more testable
