+++ go.mod
@@ -14,6 +14,7 @@ require (
 	github.com/onsi/gomega v1.6.0 // indirect
 	github.com/opencontainers/go-digest v1.0.0-rc1 // indirect
 	github.com/opencontainers/image-spec v1.0.1
+	github.com/pkg/errors v0.8.1
 	github.com/sirupsen/logrus v1.4.2
 	github.com/u-root/u-root v6.0.0+incompatible
 	github.com/vbauerster/mpb v3.4.0+incompatible
